+++ src/background-video-block/render.php
@@ -17,15 +17,6 @@ $fbks_height_tablet        = isset( $attributes['heightTablet'] ) ? (float) $att
 $fbks_height_mobile        = isset( $attributes['heightMobile'] ) ? (float) $attributes['heightMobile'] : 50;
 $fbks_overlay_color        = isset( $attributes['overlayColor'] ) ? (string) $attributes['overlayColor'] : '#000000';
 $fbks_overlay_opacity      = isset( $attributes['overlayOpacity'] ) ? (float) $attributes['overlayOpacity'] : 0.35;
-$fbks_layout_type          = (
-	isset( $attributes['layout'] ) &&
-	is_array( $attributes['layout'] ) &&
-	isset( $attributes['layout']['type'] )
-) ? (string) $attributes['layout']['type'] : '';
-$fbks_legacy_content_width = ! isset( $attributes['innerBlocksUseContentWidth'] ) || ! empty( $attributes['innerBlocksUseContentWidth'] );
-$fbks_use_content_width    = in_array( $fbks_layout_type, array( 'flow', 'default' ), true )
-	? false
-	: ( $fbks_layout_type ? true : $fbks_legacy_content_width );
 // Responsive focal point defaults (fallback to legacy single-value attributes if present).
 $fbks_object_pos_x_desktop = isset( $attributes['objectPositionXDesktop'] ) ? (float) $attributes['objectPositionXDesktop'] : ( isset( $attributes['objectPositionX'] ) ? (float) $attributes['objectPositionX'] : 50 );
 $fbks_object_pos_x_tablet  = isset( $attributes['objectPositionXTablet'] ) ? (float) $attributes['objectPositionXTablet'] : $fbks_object_pos_x_desktop;
@@ -75,6 +66,7 @@ $fbks_view_data = array(
 	'provider'       => $fbks_media_provider,
 	'url'            => $fbks_media_url,
 	'vimeoId'        => $fbks_vimeo_id,
+	'vimeoAspectRatio' => isset( $attributes['vimeoAspectRatio'] ) ? (float) $attributes['vimeoAspectRatio'] : null,
 	'loop'           => $fbks_loop,
 	'disableMobile'  => $fbks_disable_mobile,
 	'playOnHover'    => $fbks_play_on_hover,
@@ -91,8 +83,7 @@ if ( $fbks_has_media ) {
 		'--pb-bgvid-h-desktop:%1$svh;--pb-bgvid-h-tablet:%2$svh;--pb-bgvid-h-mobile:%3$svh;'
 		. '--pb-bgvid-pos-x-desktop:%4$s%%;--pb-bgvid-pos-x-tablet:%5$s%%;--pb-bgvid-pos-x-mobile:%6$s%%;'
 		. '--pb-bgvid-pos-y-desktop:%7$s%%;--pb-bgvid-pos-y-tablet:%8$s%%;--pb-bgvid-pos-y-mobile:%9$s%%;'
-		. '--pb-bgvid-pos-x:%4$s%%;--pb-bgvid-pos-y:%7$s%%;'
-		. '--pb-bgvid-overlay:%10$s;',
+		. '--pb-bgvid-pos-x:%4$s%%;--pb-bgvid-pos-y:%7$s%%;',
 		esc_attr( $fbks_height_desktop ),
 		esc_attr( $fbks_height_tablet ),
 		esc_attr( $fbks_height_mobile ),
@@ -101,17 +92,20 @@ if ( $fbks_has_media ) {
 		esc_attr( $fbks_object_pos_x_mobile ),
 		esc_attr( $fbks_object_pos_y_desktop ),
 		esc_attr( $fbks_object_pos_y_tablet ),
-		esc_attr( $fbks_object_pos_y_mobile ),
-		esc_attr( sprintf( 'rgba(0,0,0,0)' ) )
+		esc_attr( $fbks_object_pos_y_mobile )
 	);
 }
 
-$fbks_wrapper_attributes = get_block_wrapper_attributes(
-	array(
-		'class' => $fbks_has_media ? 'pb-bgvid has-media' : 'pb-bgvid',
-		'style' => $fbks_wrapper_style,
-	)
+$fbks_wrapper_args = array(
+	'class' => $fbks_has_media ? 'pb-bgvid has-media' : 'pb-bgvid',
+	'style' => $fbks_wrapper_style,
 );
+
+if ( ! empty( $attributes['disableRightClick'] ) ) {
+	$fbks_wrapper_args['data-disable-right-click'] = 'true';
+}
+
+$fbks_wrapper_attributes = get_block_wrapper_attributes( $fbks_wrapper_args );
 
 // Overlay is output as inline style because we store a hex + opacity in attributes.
 $fbks_overlay_style = sprintf(
@@ -120,48 +114,11 @@ $fbks_overlay_style = sprintf(
 	esc_attr( $fbks_overlay_opacity )
 );
 $fbks_layer_style = 'width:100%;max-width:none;margin:0;';
-$fbks_content_inner_style = '';
-$fbks_block_gap = isset( $attributes['style']['spacing']['blockGap'] ) ? $attributes['style']['spacing']['blockGap'] : null;
-$fbks_normalize_gap_value = static function( $value ) {
-	if ( ! is_string( $value ) && ! is_numeric( $value ) ) {
-		return '';
-	}
-
-	$value = (string) $value;
-	if ( '' === $value || preg_match( '%[\\\(&=}]|/\*%', $value ) ) {
-		return '';
-	}
-
-	if ( str_contains( $value, 'var:preset|spacing|' ) ) {
-		$index_to_splice = strrpos( $value, '|' ) + 1;
-		$slug            = _wp_to_kebab_case( substr( $value, $index_to_splice ) );
-		$value           = "var(--wp--preset--spacing--$slug)";
-	}
-
-	return safecss_filter_attr( explode( ';', $value )[0] );
-};
-
-if ( is_array( $fbks_block_gap ) ) {
-	$fbks_row_gap    = isset( $fbks_block_gap['top'] ) ? $fbks_normalize_gap_value( $fbks_block_gap['top'] ) : '';
-	$fbks_column_gap = isset( $fbks_block_gap['left'] ) ? $fbks_normalize_gap_value( $fbks_block_gap['left'] ) : '';
-	if ( '' !== $fbks_row_gap ) {
-		$fbks_content_inner_style .= "row-gap:$fbks_row_gap;";
-	}
-	if ( '' !== $fbks_column_gap ) {
-		$fbks_content_inner_style .= "column-gap:$fbks_column_gap;";
-	}
-} else {
-	$fbks_row_gap = $fbks_normalize_gap_value( $fbks_block_gap );
-	if ( '' !== $fbks_row_gap ) {
-		$fbks_content_inner_style .= "row-gap:$fbks_row_gap;";
-	}
-}
 
 $fbks_content_inner_classes = sprintf(
-	'pb-bgvid__content-inner is-h-%1$s is-v-%2$s %3$s',
+	'pb-bgvid__content-inner is-h-%1$s is-v-%2$s is-full-width',
 	sanitize_html_class( $fbks_items_justification ),
-	sanitize_html_class( $fbks_vertical_alignment ),
-	$fbks_use_content_width ? 'is-content-width' : 'is-full-width'
+	sanitize_html_class( $fbks_vertical_alignment )
 );
 
 ?>
@@ -173,7 +130,7 @@ $fbks_content_inner_classes = sprintf(
 	</div>
 
 	<div class="pb-bgvid__content alignfull" style="<?php echo esc_attr( $fbks_layer_style ); ?>">
-		<div class="<?php echo esc_attr( $fbks_content_inner_classes ); ?>"<?php echo '' !== $fbks_content_inner_style ? ' style="' . esc_attr( $fbks_content_inner_style ) . '"' : ''; ?>>
+		<div class="<?php echo esc_attr( $fbks_content_inner_classes ); ?>">
 			<?php
 			// Inner blocks.
 			echo wp_kses_post( $content );